+++ aiohttp_client_cache/backends/base.py
@@ -6,7 +6,8 @@ from abc import ABCMeta, abstractmethod
 from collections import UserDict
 from datetime import datetime
 from logging import getLogger
-from typing import Any, AsyncIterable, Awaitable, Callable, Iterable, Union
+from typing import Any, Callable, Union
+from collections.abc import AsyncIterable, Awaitable, Iterable
 
 from aiohttp import ClientResponse
 from aiohttp.typedefs import StrOrURL