+++ aiohttp_client_cache/backends/dynamodb.py
@@ -2,7 +2,8 @@ from __future__ import annotations
 
 from contextlib import asynccontextmanager
 from logging import getLogger
-from typing import Any, AsyncIterable
+from typing import Any
+from collections.abc import AsyncIterable
 
 import aioboto3
 from aioboto3.session import ResourceCreatorContext