+++ aiohttp_client_cache/backends/mongodb.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
-from typing import Any, AsyncIterable
+from typing import Any
+from collections.abc import AsyncIterable
 
 from motor.motor_asyncio import AsyncIOMotorClient
 from pymongo import MongoClient