+++ aiohttp_client_cache/backends/redis.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
-from typing import Any, AsyncIterable
+from typing import Any
+from collections.abc import AsyncIterable
 
 from redis.asyncio import Redis, from_url